+++ tests/testthat/test-get_NE_data.R
@@ -1,6 +1,6 @@
 test_that("get_ne_data retrieves map data correctly", {
   # Test country level
-  france_map <- get_ne_data(latlong_bbox = NULL,
+  france_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                             projected_crs = "EPSG:3857",
                             region = "France",
                             level = "country",
@@ -10,7 +10,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_equal(sf::st_crs(france_map[[1]]), sf::st_crs("EPSG:3857"))
 
   # Test continent level
-  africa_map <- get_ne_data(latlong_bbox = NULL,
+  africa_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                             projected_crs = "EPSG:3857",
                             region = "Africa",
                             level = "continent",
@@ -20,7 +20,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_equal(sf::st_crs(africa_map[[2]]), sf::st_crs("EPSG:3857"))
 
   # Test world level
-  world_map <- get_ne_data(latlong_bbox = NULL,
+  world_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                            projected_crs = "EPSG:3857",
                            region = NULL,
                            level = "world",
@@ -30,7 +30,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_equal(sf::st_crs(world_map[[1]]), sf::st_crs("EPSG:3857"))
 
   # Test sovereignty level
-  sovereignty_map <- get_ne_data(latlong_bbox = NULL,
+  sovereignty_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                                  projected_crs = "EPSG:3857",
                                  region = "United States of America",
                                  level = "sovereignty",
@@ -39,7 +39,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_s3_class(sovereignty_map[[2]], "sf")
 
   # Test geounit level
-  geounit_map <- get_ne_data(latlong_bbox = NULL,
+  geounit_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                              projected_crs = "EPSG:3857",
                              region = "Germany",
                              level = "geounit",
@@ -48,7 +48,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_s3_class(geounit_map[[1]], "sf")
 
   # Test tiny_countries
-  tiny_countries_map <- get_ne_data(latlong_bbox = NULL,
+  tiny_countries_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                                     projected_crs = "EPSG:3857",
                                     region = "France",
                                     level = "country",
@@ -57,7 +57,7 @@ test_that("get_ne_data retrieves map data correctly", {
   expect_s3_class(tiny_countries_map[[2]], "sf")
 
   # test multiple level ne_type combinations.
-  sovereignty_tiny_map <- get_ne_data(latlong_bbox = NULL,
+  sovereignty_tiny_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                                       projected_crs = "EPSG:3857",
                                       region = "United States of America",
                                       level = "sovereignty",
@@ -68,7 +68,7 @@ test_that("get_ne_data retrieves map data correctly", {
 
 test_that("get_ne_data handles edge cases", {
   # Test different ne_scale
-  france_map_small <- get_ne_data(latlong_bbox = NULL,
+  france_map_small <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                                   projected_crs = "EPSG:3857",
                                   region = "France",
                                   level = "country",
@@ -77,7 +77,7 @@ test_that("get_ne_data handles edge cases", {
   expect_s3_class(france_map_small[[2]], "sf")
 
   # Test different region
-  germany_map <- get_ne_data(latlong_bbox = NULL,
+  germany_map <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                              projected_crs = "EPSG:3857",
                              region = "Germany",
                              level = "country",
@@ -88,7 +88,7 @@ test_that("get_ne_data handles edge cases", {
 
 test_that("get_ne_data handles errors", {
   # Test invalid level
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(llatlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                            output_crs = "EPSG:3857",
                            region = "France",
                            level = "invalid",
@@ -96,31 +96,31 @@ test_that("get_ne_data handles errors", {
                            ne_scale = "medium"))
 
   # Test that when region and latlong_bbox are both null we get an error
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                               projected_crs = "EPSG:3857",
                               region = NULL,
                               level = "country",
                               ne_type = "countries",
                               ne_scale = "medium"))
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                               projected_crs = "EPSG:3857",
                               region = NULL,
                               level = "continent",
                               ne_type = "countries",
                               ne_scale = "medium"))
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                               projected_crs = "EPSG:3857",
                               region = NULL,
                               level = "sovereignty",
                               ne_type = "sovereignty",
                               ne_scale = "medium"))
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                               projected_crs = "EPSG:3857",
                               region = NULL,
                               level = "geounit",
                               ne_type = "map_units",
                               ne_scale = "medium"))
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                               projected_crs = "EPSG:3857",
                               region = NULL,
                               level = "country",
@@ -128,7 +128,7 @@ test_that("get_ne_data handles errors", {
                               ne_scale = "medium"))
 
   # Test for a valid result when ne_type is NULL
-  no_ne_type <- get_ne_data(latlong_bbox = NULL,
+  no_ne_type <- get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                             projected_crs = "EPSG:3857",
                             region = "France",
                             level = "country",
@@ -137,7 +137,7 @@ test_that("get_ne_data handles errors", {
   expect_s3_class(no_ne_type[[1]], "sf")
 
   # Test for error when level is null
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                            projected_crs = "EPSG:3857",
                            region = "Germany",
                            level = NULL,
@@ -145,7 +145,7 @@ test_that("get_ne_data handles errors", {
                            ne_scale = "medium"))
 
   # Test for error when output_crs is null
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                            projected_crs = NULL,
                            region = "Germany",
                            level = "country",
@@ -153,7 +153,7 @@ test_that("get_ne_data handles errors", {
                            ne_scale = "medium"))
 
   # Test for error when ne_scale is null
-  expect_error(get_ne_data(latlong_bbox = NULL,
+  expect_error(get_ne_data(latlong_bbox <- c(xmin = 10, ymin = 40, xmax = 12, ymax = 42),
                            projected_crs = "EPSG:3857",
                            region = "Germany",
                            level = "country",